finished

+++ react-app/src/components/projects/DeleteProject.js
@@ -3,6 +3,7 @@ import {  useDispatch } from 'react-redux';
 import {useHistory} from 'react-router-dom';
 import { Modal } from '../context/Modal';
 import { DeleteProjects } from '../../store/projects';
+import { DeleteRelatedTask } from '../../store/tasks';
 import './DeleteProjectModal.css'
 
 function DeleteProjectModal({id, onClose}){
@@ -10,10 +11,14 @@ function DeleteProjectModal({id, onClose}){
     const history = useHistory();
     async function onDelete(e){
         e.preventDefault();
-       const deleted =  await dispatch(DeleteProjects(id))
-       if(!deleted){
-           history.push('/')
-       }
+        const deletedTask = await dispatch(DeleteRelatedTask(id))
+        if(!deletedTask){
+            const deleted =  await dispatch(DeleteProjects(id))
+            if(!deleted){
+                history.push('/')
+            }
+        }
+
     }
     return (
         <>
